feat(mobilitybox): adds location parameter support for search_by_name function
This change reflects the new option to supply a location to the search_by_name endpoint. It is
optional. This is a breaking change because we use an object as the main parameter for this function
now (also known as "named parameters").

BREAKING CHANGE: station_by_name now uses named parameters (an object) as the first argument. This
allows you to freely change the order of arguments.